requirements.txt
altair==5.5.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
streamlit==1.32.2
rich>=10.14.0,<14
gspread==6.2.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-docx==1.1.0
pymupdf==1.23.21
joblib==1.3.2
